refactor: update import paths for InMemoryCache and remove redundant checks in workflows
