+++ .github/workflows/Push.yml
@@ -28,10 +28,6 @@ jobs:
     steps:
       - name: Checkout code
         uses: actions/checkout@v2
-      - name: Check if package.json changed
-        run: |
-          CHANGED=$(git diff HEAD~1 HEAD --name-only | grep "^package.json" || true)
-          if [ -z "$CHANGED" ]; then echo "No changes in package.json, skipping Publish_To_NPM."; exit 0; fi
       - name: Install dependencies
         run: npm ci
       - name: Build package
@@ -51,10 +47,6 @@ jobs:
     steps:
       - name: Checkout code
         uses: actions/checkout@v2
-      - name: Check if package.json changed
-        run: |
-          CHANGED=$(git diff HEAD~1 HEAD --name-only | grep "^package.json" || true)
-          if [ -z "$CHANGED" ]; then echo "No changes in package.json, skipping Publish_To_Github."; exit 0; fi
       - name: Install dependencies
         run: npm install
       - name: Build package
@@ -78,10 +70,6 @@ jobs:
     steps:
       - name: Checkout code
         uses: actions/checkout@v2
-      - name: Check if Document directory changed
-        run: |
-          CHANGED=$(git diff HEAD~1 HEAD --name-only | grep "^Document/" || true)
-          if [ -z "$CHANGED" ]; then echo "No changes in Document directory, skipping Send_To_Server."; exit 0; fi
       - name: Setup SSH
         uses: webfactory/ssh-agent@v0.5.4
         with:
