+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "axiodb",
-  "version": "2.19.65",
+  "version": "2.19.66",
   "description": "A blazing-fast, lightweight, and scalable nodejs package based DBMS for modern application. Supports schemas, encryption, and advanced query capabilities.",
   "main": "./lib/config/DB.js",
   "types": "./lib/config/DB.d.ts",
