+++ source/Services/CRUD Operation/Delete.operation.ts
@@ -13,7 +13,7 @@ import FolderManager from "../../engine/Filesystem/FolderManager";
 // Import All Utility
 import Searcher from "../../utility/Searcher.utils";
 import Sorting from "../../utility/SortData.utils";
-import InMemoryCache from "../../cache/cache.operation";
+import InMemoryCache from "../../caching/cache.operation";
 import { General } from "../../config/Keys/Keys";
 import ReaderWithWorker from "../../utility/BufferLoaderWithWorker.utils";
 /**
